+++ client/src/features/hierarchy/ChildCustomNode.tsx
@@ -6,13 +6,20 @@ import config from '../../config/config'
 import { stripYaleIdPrefix } from '../../lib/parse/data/helper'
 import RecordLink from '../common/RecordLink'
 import StyledHierarchyButton from '../../styles/features/hierarchy/HierarchyButton'
+import { useAppDispatch } from '../../app/hooks'
+import { addInitialState } from '../../redux/slices/hierarchyVisualizationSlice'
 
 interface IProps {
   entityId: string
 }
 
 const ChildCustomNode: React.FC<IProps> = ({ entityId }) => {
+  const dispatch = useAppDispatch()
   const uriToRetrieve = stripYaleIdPrefix(entityId)
+
+  const handleHierarchyChange = (p: Array<string>): void => {
+    dispatch(addInitialState({ origin: entityId, parents: p, children: [] }))
+  }
 
   const { data, isSuccess } = useGetItemQuery({
     uri: uriToRetrieve,
@@ -20,9 +27,11 @@ const ChildCustomNode: React.FC<IProps> = ({ entityId }) => {
   })
 
   let name = ''
+  let parents: Array<string> = []
   if (data && isSuccess) {
     const entity = new EntityParser(data)
     name = entity.getPrimaryName(config.dc.langen)
+    parents = entity.getPartOf()
   }
 
   return (
@@ -33,6 +42,7 @@ const ChildCustomNode: React.FC<IProps> = ({ entityId }) => {
           type="button"
           rotate="-90"
           className="childButton"
+          onClick={handleHierarchyChange(parents)}
           aria-label={`View the hierarchy for ${name}`}
         >
           <i className="bi bi-diagram-2 fs-5" />